Fix comparison of enum to -1 compiler warning
    xlnk.c:753:34: warning: comparison of constant -1 with expression of
    type 'xasm_constant_type' (aka 'enum tag_xasm_constant_type') is always
    false [-Wtautological-constant-out-of-range-compare]
            if ((lhs_result.type == -1) || (rhs_result.type == -1)) {

+++ script.c
@@ -119,7 +119,7 @@ static int is_valid_command_arg(xlnk_command_type type, const char *candidate_ar
         { XLNK_PAD_COMMAND,      pad_args }
     };
     /* Find arg array for command */
-    for (i=0; ok_args[i].type != -1; i++) {
+    for (i=0; (int)ok_args[i].type != -1; i++) {
         if (ok_args[i].type == type) {
             /* Now go through array of valid args for command */
             args = ok_args[i].args;
